Added Bullet Spawing (No moving)
Changed how the drawing of spriteObjects are drawn. Is now Recersive.

Fixed GetRotation, now returns the proper value

Added get methods to get the foward vector of a matrix and the ability to get the positon of an offset of the foward vector

Added a number of constructers for creating gameObjects

Bullets now spawn at the tip of the tank turret when space is pressed but do not move yet